Moved user management endpoints from user service to security service
