+++ Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/security/requests/AddUserRequest.kt
@@ -1,4 +1,4 @@
-package com.savannasolutions.SmartContractVerifierServer.user.requests
+package com.savannasolutions.SmartContractVerifierServer.security.requests
 
 import com.fasterxml.jackson.annotation.JsonProperty
