+++ Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/security/responses/AddUserResponse.kt
@@ -1,4 +1,4 @@
-package com.savannasolutions.SmartContractVerifierServer.user.responses
+package com.savannasolutions.SmartContractVerifierServer.security.responses
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.savannasolutions.SmartContractVerifierServer.common.ResponseStatus
