--- Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/user/requests/GetNonceRequest.kt
@@ -1,5 +1,0 @@
-package com.savannasolutions.SmartContractVerifierServer.user.requests
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class GetNonceRequest(@JsonProperty("UserID") val UserID: String,)
