+++ Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/user/services/UserService.kt
@@ -8,34 +8,19 @@ import com.savannasolutions.SmartContractVerifierServer.negotiation.repositories
 import com.savannasolutions.SmartContractVerifierServer.negotiation.repositories.ConditionsRepository
 import com.savannasolutions.SmartContractVerifierServer.user.models.User
 import com.savannasolutions.SmartContractVerifierServer.user.repositories.UserRepository
-import com.savannasolutions.SmartContractVerifierServer.user.requests.AddUserRequest
-import com.savannasolutions.SmartContractVerifierServer.user.requests.GetNonceRequest
+import com.savannasolutions.SmartContractVerifierServer.security.requests.AddUserRequest
 import com.savannasolutions.SmartContractVerifierServer.user.requests.RetrieveUserAgreementsRequest
-import com.savannasolutions.SmartContractVerifierServer.user.requests.UserExistsRequest
-import com.savannasolutions.SmartContractVerifierServer.user.responses.AddUserResponse
-import com.savannasolutions.SmartContractVerifierServer.user.responses.GetNonceResponse
+import com.savannasolutions.SmartContractVerifierServer.security.requests.UserExistsRequest
+import com.savannasolutions.SmartContractVerifierServer.security.responses.AddUserResponse
 import com.savannasolutions.SmartContractVerifierServer.user.responses.RetrieveUserAgreementsResponse
-import com.savannasolutions.SmartContractVerifierServer.user.responses.UserExistsResponse
+import com.savannasolutions.SmartContractVerifierServer.security.responses.UserExistsResponse
 import org.springframework.stereotype.Service
-import java.util.concurrent.ThreadLocalRandom
 import kotlin.collections.ArrayList
 
 @Service
 class UserService(  val userRepository: UserRepository,
                     val agreementsRepository: AgreementsRepository,
                     val conditionsRepository: ConditionsRepository) {
-
-    fun addUser(addUserRequest: AddUserRequest): AddUserResponse {
-        val wID = addUserRequest.WalletID
-        if(userRepository.existsById(wID))
-            return AddUserResponse(status = ResponseStatus.FAILED)
-
-        val user = User(publicWalletID = addUserRequest.WalletID, alias = addUserRequest.Alias)
-
-        userRepository.save(user)
-
-        return  AddUserResponse(status = ResponseStatus.SUCCESSFUL)
-    }
 
 
     fun retrieveUserAgreements(retrieveUserAgreementsRequest: RetrieveUserAgreementsRequest): RetrieveUserAgreementsResponse {
@@ -99,26 +84,4 @@ class UserService(  val userRepository: UserRepository,
         return RetrieveUserAgreementsResponse(list,ResponseStatus.SUCCESSFUL)
     }
 
-    fun userExists(userExistsRequest: UserExistsRequest): UserExistsResponse
-    {
-        if(userExistsRequest.walletID.isEmpty())
-            return UserExistsResponse(status = ResponseStatus.FAILED)
-
-        return if(userRepository.existsById(userExistsRequest.walletID))
-            UserExistsResponse(true, ResponseStatus.SUCCESSFUL)
-        else
-            UserExistsResponse(status = ResponseStatus.SUCCESSFUL)
-    }
-
-
-    fun getNonce(getNonceRequest: GetNonceRequest): GetNonceResponse{
-        if (userRepository.existsById(getNonceRequest.UserID)) {
-            val user = userRepository.getById(getNonceRequest.UserID)
-            val nonce = ThreadLocalRandom.current().nextLong(1000000000, 9999999999)
-            user.nonce = nonce
-            userRepository.save(user)
-            return GetNonceResponse(nonce, ResponseStatus.SUCCESSFUL)
-        }
-        return GetNonceResponse(0, ResponseStatus.FAILED)
-    }
 }
