+++ Smart-Contract-Verifier-Server/src/test/kotlin/com/savannasolutions/SmartContractVerifierServer/user/services/UserServiceTest.kt
@@ -7,9 +7,8 @@ import com.savannasolutions.SmartContractVerifierServer.negotiation.repositories
 import com.savannasolutions.SmartContractVerifierServer.user.models.User
 import com.savannasolutions.SmartContractVerifierServer.user.repositories.UserRepository
 import com.savannasolutions.SmartContractVerifierServer.user.requests.RetrieveUserAgreementsRequest
-import com.savannasolutions.SmartContractVerifierServer.user.requests.UserExistsRequest
+import com.savannasolutions.SmartContractVerifierServer.security.requests.UserExistsRequest
 import org.junit.Test
-import org.mockito.kotlin.any
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.whenever
 import java.util.*
